Removed CommandWrapLauncher. Moved all command overrides to commandoverride.py.
